+++ packages/server-core/src/agents/AgentManager.ts
@@ -1,6 +1,6 @@
 import Agent from './Agent'
 import { projectId, AGENT_WEBSERVER_PORT_RANGE, pluginManager } from '@magickml/engine'
-import { app } from './app'
+import { app } from '../app'
 
 // if the user is running the app locally, sort by their project id
 // this way users can use our demo database without seeing each other's stuff
