+++ packages/server-core/src/config/validators.ts
