+++ packages/server-core/src/helpers/buildMagickInterface.ts
@@ -1,8 +1,8 @@
-import { runSpell } from './utils/runSpell'
-import { API_ROOT_URL, pluginManager } from '@magickml/engine'
-import { app } from './app'
+import { runSpell } from '../utils/runSpell'
+import { API_ROOT_URL } from '@magickml/engine'
+import { app } from '../app'
 
-export const buildMagickInterface = (overrides: Record<string, Function> = {}) => {
+export const buildMagickInterface = (overrides: Record<string, any> = {}) => {
   const env = {
     API_ROOT_URL,
   }
