+++ packages/server-core/src/hooks/index.ts
@@ -1,6 +1,6 @@
 // For more information about this file see https://dove.feathersjs.com/guides/cli/log-error.html
 import type { HookContext, NextFunction } from '../declarations'
-import { logger } from '../logger'
+import { logger } from '../utils/logger'
 
 export const logError = async (context: HookContext, next: NextFunction) => {
   try {
