+++ packages/server-core/src/index.ts
@@ -1,23 +1,23 @@
-export * from './Agent'
-export * from './AgentManager'
-export * from './apis'
+export * from './agents/Agent'
+export * from './agents/AgentManager'
+export * from './api/apis'
 export * from './app'
-export * from './buildMagickInterface'
-export * from './configuration'
+export * from './helpers/buildMagickInterface'
+export * from './config/configuration'
 export * from './declarations'
-export * from './fileServer'
-export * from './googleSpeechToText'
-export * from './googleTextToSpeech'
+export * from './servers/fileServer'
+export * from './servers/googleSpeechToText'
+export * from './servers/googleTextToSpeech'
 export * from './hooks'
-export * from './logger'
+export * from './utils/logger'
 export * from './dbClient'
-export * from './queryGoogleSearch'
-export * from './ServerError'
+export * from './servers/queryGoogleSearch'
+export * from './utils/ServerError'
 export * from './services'
-export * from './sockets'
+export * from './sockets/sockets'
 export * from './spells'
-export * from './tiktalknet'
-export * from './types'
+export * from './servers/tiktalknet'
+export * from './config/types'
 export * from './utils'
-export * from './validators'
+export * from './config/validators'
