+++ packages/server-core/src/servers/googleSpeechToText.ts
