+++ packages/server-core/src/services/agents/agents.schema.ts
@@ -4,7 +4,7 @@ import { Type, getDataValidator, getValidator, querySyntax } from '@feathersjs/t
 import type { Static } from '@feathersjs/typebox'
 
 import type { HookContext } from '../../declarations'
-import { dataValidator, queryValidator } from '../../validators'
+import { dataValidator, queryValidator } from '../../config/validators'
 
 // Main data model schema
 export const agentSchema = Type.Object(
