+++ packages/server-core/src/services/spells/spells.schema.ts
@@ -3,7 +3,7 @@ import { Type, getDataValidator, getValidator, querySyntax } from '@feathersjs/t
 import type { Static } from '@feathersjs/typebox'
 
 import type { HookContext } from '../../declarations'
-import { dataValidator, queryValidator } from '../../../../../packages/server-core/src/validators'
+import { dataValidator, queryValidator } from '../../config/validators'
 
 // Main data model schema
 export const spellSchema = Type.Object(
