+++ packages/server-core/src/services/spells/spells.ts
@@ -84,7 +84,7 @@ export const spell = (app: Application) => {
         schemaHooks.validateData(spellPatchValidator),
         schemaHooks.resolveData(spellPatchResolver),
         handleJSONFieldsUpdate(spellJsonFields),
-        // updateSpellInManager
+        updateSpellInManager
       ],
       update: [handleJSONFieldsUpdate(spellJsonFields)],
       remove: [],
